+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- * add to here... -->
 ### Changed
 * Update form-data lib due to https://github.com/advisories/GHSA-fjxv-7rqg-78g4
+* Fixed `targetLang` parameter in `DeepLClient::rephraseText` to be an optional `string` parameter.
 
 ## [1.19.0] - 2025-06-26
 ### Added